Fix populating languages table
This didn't work as db/schema.sql would insert ('', '(unknown)') for
empty/unknown values, but on startup it would check if the total rows in
the languages table was 0, which was never true.

Since collecting language data isn't enabled by default and most people
aren't interested in it, I guess no one noticed until now.

Change it to always load the file with "on conflict", so it also
processes updates. It's okay to keep "deleted" entries around.

Note: in the future we may want to do "on-demand" insertion like for
locations. There are 7086 languages, but on goatcounter.com there are
only 335 languages in use.

Fixes #808

Co-authored-by: Xing Ji <xinji@redhotcg.com>