Migrate to log/slog, add -json flag for JSON logs
Now that slog is in stdlib, we can use that instead of my own zlog.

Add a little wrapper in in the log package to preserve some semantics of
the old zlog. In particular the ability to turn logging on/off per
"module" is something I find very useful.

As a bonus, it's now also easy to add JSON logging, so add that as well.

Fixes #311