Don't push every main commit to DockerHub
Don't think that's necessarily all that useful. The idea was that people
can easily get some newer version if they want it, but building your own
image is pretty quick and easy, and I can still push individual commits
manually if it's useful.

Can re-enable if there's a lot of demand for it.