Fix merging of multiple paths when more than once path has entries for the same hour
This would error with "ON CONFLICT DO UPDATE command cannot affect row a
second time", as it tried to update the same destination row more than
once, which PostgreSQL doesn't support. So group them first with a CTE.

For hit_stats this is rather tricky as we need to join the "stats"
array, which especially in SQLite is not so easy. So just do that
separately.

Fixes #814