Allow fetching GeoIP updates, refactor GeoIP code
I've long since had to fork the upstream maxminddb and geoip packages as
my PRs weren't merged, so just "fork" the lot to the geoip package here.
Had to do some minor updates to geoip2 package to make it work with
latest maxminddb. Also don't depend on testify.

---

Integrate automatic updates, which is the main reason I started working
on this in the first place. You can now do:

    -geodb to/file.mmdb
    -geodb maxmind:account:license_key

And it will automatically download goatcounter-data/auto.mmdb, or update
it if it's older than 7 days.

Currently it will do this on startup only; too much effort to allow
in-place updates.

Fixes #818