Automatically detect cookie secure and sameSite attributes
Previously this was set at startup depending on which flags were used,
which was a bit annoying and could lead to confusing situations. Most
proxies seem to set X-Forwarded-Proto in my testing, so we can use that
for detecting the secure attribute.

Also replace the zhttp.CookieSameSite global with a
zhttp.CookieSameSiteHelper(), so it can be set per-request and only if
the site has AllowEmbed enabled.

This removes the need for -tls=proxy, which is now just the same as
-tls=none.