Store bot pageviews in new "bots" table
Useful for debugging/answering questions, especially since for most
people none of this is recorded any more. And this way it doesn't need
to scan all of the hits table to delete old bot views, which is the most
expensive query in terms of total_exec_time (by more than double).