Use "= any(..)" instead of "in (..)" on PostgreSQL
The upshot of this is that it will properly group these queries in
pg_stat_statements since it's always e.g. "= any($3)" instead of "in
($3, $4)" and "in ($3, $4, $5)", etc.

SQLite doesn't support it, so need to do a bit of hackery to make it
work there. But it's not too bad.