Actually run the PostgreSQL tests in CI.
It just tested against SQLite instead of PostgreSQL as it should be
-tags=testpg, not -tags=pgsql – no idea how I came up with pgsql, do'h.
Didn't notice before as I usually run the tests with SQLite locally and
rely on the CI for PostgreSQL.

Also fix one test that was failing: the DELETE should ways run, not just
for SQLite. appending the loadPathIDs is needed for correct SQLite
behaviour.

The other fixes are to make sure it works with CGO_ENABLED=0. Also run
it like that in the CI; it's okay to lose the race detector as the
differences between SQLite and PostgreSQL are fairly minimal.

And finally, set the default PG* variables so it works with the
PostgreSQL container from compose.yaml.