Replace "↕" with a slider in the header to scale the charts
This is much easier and nicer to use. I don't know why I didn't think of this
before.

TODO: on e.g.
https://stats.arp242.net/?hl-period=year&group=day&period-start=2024-07-03&period-end=2025-07-03&filter=&scale=134

the top is at 16,658 – which is one big spike. To get something useful
we need something like 70 or lower, which is impossible to select with
the mouse: need to use the keyboard.

Maybe the number after the slider should be an <input>.