Automatically detect WebSocket support
Reload with ?no-websocket=1 if Websockets don't appear to work; if this
happens three times in a row then just disable Websockets support on the
server.

The advantage is that you no longer need to use "serve -websocket". I
added Websockets in 2.2.0, but some people reported issues (#581, #583)
so made it optional as a quick fix in 2.3.0. I suspect many people are
missing out on this because they don't know they have to set the flag.