Add JSON export
The old CSV export works only if you have collection of individual
pageviews enabled in the settings, which is disabled by default now.

So directly export the data and stats tables, and insert them directly.

TODO: kind of works, but needs some work to not be ugly and fragile.
Ideally a lot of this can/should be merged with how cron works.