logscan: accept port in $host and $url
Apparently some programs log with the port in there. No reason to not
accept that.

Remove $host from the bunny format; that's already set from $url which
will override $host, and the "host" is a numeric "pull zone ID" so it's
not useful in the first place. Kind of confusing to have it in there.

Fixes #848