More reliable reloading of the dashboard
When changing the widget settings or filtering data the dashboard needs
to be reloaded with new data. Sometimes some widgets only displayed
"Loading…"

The problem is that the main dashboard request spawns a few goroutines
in the background that will return data over the websocket when ready.
The websockets could return data before the main dashboard request
finishes, so what happened was:

- Replace HTML from websocket.
- Replace all of the widgets on success callback from main ajax request
- End result is that the chart is just set to "Loading…"

Solution is to not replace all widgets from the AJAX request but just
hose that won't get replaced. This is also less janky visually.