Support is:event and is:pageview in the path filter
Intentionally not documented for now until I completely figure out the
filter syntax and build some reasonably nice feedback in the UI. That's
#194

However, I'm quite sure that I want to support is:event and is:pageview.
This is quick and easy to add, and solves some use cases for people.

Fixes #842