Use filter table if it selects many paths
Turns out that running "where path_id in (1 million paths)" is really
slow, both for "in (..)" and arrays. It spends minutes just in the
planning stage. Looking with gdb at what PostgreSQL is doing, almost all
time is spent in deconstruct_array(). I didn't spend too much time
investigating, but it seems this just doesn't scale well.

So use a join if lots of paths are selected. This is a bit slower, but
not too much. This also solves a limitation on SQLite which doesn't have
arrays and the number of paths is limited to SQLITE_MAX_VARIABLE_NUMBER
(32,766).

Because we don't want to run the search on the paths tables several
times, use an intermediate "filters" table.

TODO:

- Some places also exclude some paths; see if we can integrate that in
  the PathFilter too.

- We can optimize queries that filter almost all paths, especially if
  there are many. That is: run "where path_id not in (5 paths)" instead
  of "where path_id in (1 million paths)"

Fixes #842