Expand filter syntax
Right now the search field is just a simple "where path like %$1% or
title like %$1%". This expands the filter syntax with a few keywords:

  at:start, at:end         match only at the start or end of path or title
  in:path, in:title        match only path or title rather than both
  is:event, is:pageview    show only events or pageviews
  :not                     negate match

This is pretty basic, but I think this should cover most things. Can be
expanded based on demand, although one difficulty is making sure the
highlights on the frontend are identical to the matching on the backend.

Fixes #194