Add option to disable the JS-based datepicker
Somewhat inspired by https://dbushell.com/2025/11/10/pikaday and
https://pikaday.dbushell.com

Personally, I don't entirely agree with that article, and find that the
JS-based datepicker often gives a better UX than the native one, at
least on my Linux system.

Selecting the month/year is easier, it generally looks better, and it
also displays as "03/12/2025" which is "correct" in the sense that this
is what en_IE specifies as the date format, but it's not "correct" in
the sense of what I want. Customizing that on the locale level is not
straight-forward. People's preferences are complex.

And from what I know, the accessibility of native datepickers is often
not all that great either.

But anyway, adding an option for it is easy and *in principle* I don't
disagree that always using the native datepicker is better. It's just
that in practice it's often not, IMHO anyway.

On mobile systems it never used the JS datepicker, because there the
native ones practically always do give a better experience.