Remove null from refs.ref_scheme
The original meaning of ref_scheme being NULL was that there is no ref,
back when everything was directly stored in the ref_counts table without
an intermediate refs table. But since being stored in their own table it
doesn't really make much sense; it was a mistake to carry that over.
This causing issues with some other changes I want to do because null
equality is a pain.