Fix some display issues with the custom domain help
The help text would immediately show as "verified" if the custom domain
doesn't forward to GoatCounter (yet). acme.Make() returned nil (no
error) so it would run UpdateCnameSetupAt(), which is not really
correct: that should only be done if acme.Make() fully succeeds. So add
a bool return for it.

Also don't update the field every time it checks for renewals: only set
it on the first creation. And unset it again if the forwarding breaks.

Also:

- Clear cname and cname_setup_at on blanking the field so the help text
  shows correctly after blanking.

- Correctly display domain in help text. Looks like this has been broken
  ever since the translation tags were added in 2021 (a04f92e5) – guess
  no one noticed because it showed as "verified" if it wasn't set up
  yet.

- Only get server IP once in acme.Make(); I guess the IP can change, but
  this is exceedingly rare, and can just restart goatcounter to fix.