Make default timerange for API identical to the dashboard
When no start parameter is given, the API would use exactly 7 days ago,
leaving the time alone. Whereas the dashboard fetches all data from the
start of the day. That makes more sense, and it's confusing if they
differ.

Also use the end of the day for end. Shouldn't really matter, but it's
what the dashboard does so make it identical.