Optimise CSP middleware
This gets run on every request, and was relatively inefficient with a
fair amount of allocations, so optimize things a bit to allocate less.

After I did the optimisation, I realized that most invocations are for
/count, where it's quite useless. So skip setting the CSP there.
Probably just skipping that would be enough, but ah well – the optimized
code isn't that much uglier than the original so just keep it.