Don't update languages on every restart
The "do update set name = excluded.name" causes the entire table to be
replaced on every restart, which causes autovacuum to run very
frequently on goatcounter.com. It's a small table and fairly harmless,
but it's clobbering the logs and been bugging me.

We can update by just doing "delete from languages" in a migration every
year or two. It doesn't change that often.